SoraDisconnectReason をファイルに切り出す
